+++ ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: okochulo <okochulo@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/12 17:32:35 by okochulo          #+#    #+#             */
-/*   Updated: 2025/05/21 17:00:26 by okochulo         ###   ########.fr       */
+/*   Updated: 2025/05/23 13:04:48 by okochulo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (size > 0)
